+++ src/utils/employees/index.ts
@@ -1,1 +1,2 @@
 export * from './employeeMedical';
+export * from './employeeEmployment';